+++ tests/E2E/ResolveOffsetSpecE2ETest.php
@@ -7,7 +7,7 @@ namespace CrazyGoat\RabbitStream\Tests\E2E;
 use CrazyGoat\RabbitStream\Request\CreateRequestV1;
 use CrazyGoat\RabbitStream\Request\ExchangeCommandVersionsRequestV1;
 use CrazyGoat\RabbitStream\Request\OpenRequestV1;
-use CrazyGoat\RabbitStream\Request\PeerPropertiesToStreamBufferV1;
+use CrazyGoat\RabbitStream\Request\PeerPropertiesRequestV1;
 use CrazyGoat\RabbitStream\Request\ResolveOffsetSpecRequestV1;
 use CrazyGoat\RabbitStream\Request\SaslAuthenticateRequestV1;
 use CrazyGoat\RabbitStream\Request\SaslHandshakeRequestV1;
@@ -42,7 +42,7 @@ class ResolveOffsetSpecE2ETest extends TestCase
             $connection = new StreamConnection(self::$host, self::$port);
             $connection->connect();
 
-            $connection->sendMessage(new PeerPropertiesToStreamBufferV1());
+            $connection->sendMessage(new PeerPropertiesRequestV1());
             $connection->readMessage();
 
             $connection->sendMessage(new SaslHandshakeRequestV1());
@@ -97,7 +97,7 @@ class ResolveOffsetSpecE2ETest extends TestCase
         $connection = new StreamConnection(self::$host, self::$port);
         $connection->connect();
 
-        $connection->sendMessage(new PeerPropertiesToStreamBufferV1());
+        $connection->sendMessage(new PeerPropertiesRequestV1());
         $connection->readMessage();
 
         $connection->sendMessage(new SaslHandshakeRequestV1());